+++ app/src/androidTest/java/com/chatApp/chatapp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.insurance.chatapp
+package com.chatApp.chatapp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
