+++ app/src/main/java/com/chatApp/chatapp/MainActivity.kt
@@ -1,10 +1,9 @@
-package com.insurance.chatapp
+package com.chatApp.chatapp
 
-import android.app.UiModeManager
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatDelegate
-import com.insurance.chatapp.databinding.ActivityMainBinding
+import com.chatApp.chatapp.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
