+++ app/src/main/res/layout/activity_main.xml
@@ -19,7 +19,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/top_user_fragment"
-        android:name="com.insurance.chatapp.TopUserFragment"
+        android:name="com.chatApp.chatapp.TopUserFragment"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dimen_0_dp"
         android:layout_weight="1"
@@ -33,7 +33,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/bottom_user_fragment"
-        android:name="com.insurance.chatapp.BottomUserFragment"
+        android:name="com.chatApp.chatapp.BottomUserFragment"
         android:layout_width="match_parent"
         android:layout_height="@dimen/dimen_0_dp"
         android:layout_weight="1"
